Implemented BMI160 lib NOT WORKING

+++ include/sensor_data.h
@@ -1,6 +1,6 @@
 #pragma once
 
 struct SensorDataIMU {
-    long IMU[3]; // XYZ
-    long gyro[3]; // XYZ
+    int16_t IMU[3]; // XYZ
+    int16_t gyro[3]; // XYZ
 };
